Remove note about authenticating as user from README.md
We have a vestigial reference to authenticating *as a user* vs *as a machine* in regard to a prompt that the Auth0 CLI presents the user with when a user uses an `auth0 login` command. However, this doesn't happen when we use the `--scopes` flag that is in our README.md instructions. This PR removes the reference that might confuse users who are asked to make sure they authenticate as a user, but are never prompted to do so in their terminal.